verified  read write

+++ src/pio_func.h
@@ -33,7 +33,7 @@ struct pio_struct_Lihan{
 };
 
 
-static uint8_t mk_cmd_buf_lihan(uint8_t *pdst, uint8_t opcode, uint16_t addr) ;
+static uint8_t mk_cmd_buf_lihan(uint32_t *pdst, uint8_t opcode, uint16_t addr) ;
 void signal_handler(int sig);
 
 
@@ -41,10 +41,10 @@ void wiznet_spi_pio_read_byte( uint8_t op_code, uint16_t AddrSel, uint8_t *rx, u
 void wiznet_spi_pio_write_byte( uint8_t op_code, uint16_t AddrSel, uint8_t *tx, uint16_t tx_length);
 
 void pio_read_byte(struct pio_struct_Lihan *pioStruct, uint8_t op_code, uint16_t AddrSel, uint8_t *rx, uint16_t rx_length);
-void pio_write_byte(struct pio_struct_Lihan *pioStruct, uint8_t op_code, uint16_t AddrSel, uint8_t *tx, uint16_t tx_length);
+void pio_write_byte(struct pio_struct_Lihan *pioStruct, uint8_t op_code, uint16_t AddrSel, uint32_t *tx, uint16_t tx_length);
 
-static uint16_t mk_cmd_buf_include_data(uint8_t *outbuf,
-                                        uint8_t *databuf, 
+static uint16_t mk_cmd_buf_include_data(uint32_t *outbuf,
+                                        uint32_t *databuf, 
                                         uint8_t opcode, 
                                         uint16_t rag_addr,  
                                         uint16_t len_byte) ;
